Payslips get Bonuses in Earnings
Payslips earnings now show bonuses brought in from those
assigned to the employee at the time that the payroll is
processed as well as work hours. These are displayed on
the payslip's earnings.

Additionally, I added a migration to change the booleans
in the tables to be not null and default to false.  This
way booleans will always be either true or false.  I
believe this is better behavior than a tri-state booleans
(true, false, and nil).

I also noticed that employees weren't displaying errors
on an update/edit, so I made a small change to show what
errors exist on the object on the form when submitting an
edit. A better visual display is needed, but this works
for now.

Note: some questions still exist about how work hours and
percentage-based bonuses work with the wage field in the
employee.  I'm not sure yet how to properly compute pay
based on the records we have.  We will need to ask the
customer about this.