Payslips are now unique for periods
Payslips now use Period's month and year in its attributes
so that it can marshal a period model for comparisons. Previously
payslips were keeping track of their own Dates for start and end.

As a result of this change, two paychecks cannot be recreated for
the same period.  If this is attempted, the previous paycheck will
be found and re-processed (deleting all child objects and
recreating them from the available data).

Note: Existing payslips and earnings may need to be cleared after
the migration (either manually, or by a rails drop/create/migrate
cycle).