Mid-Month Advance Payslips
An administrator can run a mid-month payslip and see how
much advance should be received by the employee.  When
fully processed, an appropriate charge is created for the
employee, which is turned into a deduction on the paycheck.

Note: Some of the rules are a bit fuzzy, but they should
be easy to change once we get details.