[Work Hours] Can now specify work loans
Work Hours can have a department attached to them, indicating
that the work hours record should be charged to a different
department than the employee they belong to.

Note: Only single day records can be work loaned.  In the
future, we might need to ability to work loan multiple segments
of a 8 hour day to different departments.  This cannot yet be
done.