Wages Information
Add the wages table from the old EPS system.  This adds the
migration, and the data is loaded via seeds.rb. The new model
_wage_ also is a value object to represent the data contained
in the table.

The employee interface is augmented with wage.  Previously,
this was just a normal field, but now it will pull the wage
information from the wages table based on the category and
echelon selected for the employee.  If the echelon is 'g',
then the text field for wage must be used.

The Wage model contains methods to convert the names/symbols
of the employee category and echelon enums into the format
needed to lookup in the table.  This is a bit awkward.

Also, this updates some of the tests for the new structure
of the WorkHours.total_hours hash response in payslip_test

**IMPORTANT NOTE**

_The repository doesn't actually contain the source data for
the wages table. Therefore, you need to make sure you get
this file from Brian or Rick_

After you get the file (wages.sql), put it into your ./db/
directory and you can load the data using `rails db:seed`

**END IMPORTANT NOTE**