Audited Gem for Audit Trails
Models can now be enabled for audit tracking. This has been
added to the Employee model only as a start.  Also its
history now appears in the "Activity Feed" on the home page
as a proof of concept. Other models can be added as we
decide to add them as the project progresses.

Note: There are two gems that seem to offer this same
functionality for rails projects. I've chosen `audited` after
just a few minutes of searching. The features seem fairly
identical, but I'm not sure yet that I've made the right
choice.  You can read more about the options at the following
links:

[Audited, the implemented option](https://github.com/collectiveidea/audited)
[Paper Trail](https://github.com/airblade/paper_trail)
[A comparison of the two](https://revs.runtime-revolution.com/the-battle-for-auditing-and-versioning-in-rails-audited-vs-paper-trail-17ad0011ecd9)