Union Dues and Amical now deductions on payslip
The items in the employee model that don't exist as
charges that need to be deducted on a paycheck can be
asked for by calling 'deductable_expenses()' on an
Employee instance.

Also, the `amount` field on the deduction model was
changed to decimal from integer.  There are no amounts
of CFA less than 1, but it seems better to store all
of the amounts with fractional numbers (as necessary)
and handle properly displaying these values in the
views through the localization modules. Hopefully this
is better behavior. I also added tests to exhibit this
behavior.

I also cleaned up some nil checks in the i18n of the
employee page. human_attribute_name() is not nil safe.