Basic Reports
Implemented two basic reports using the `dossier` gem.  This gem
allows you to define a sql query in a report class and most other
things 'just work'.  Kind of.

Export as CSV and XLS work. However, the exported data is more
raw (it skips the formatting of the data on export so any
formatting is lost and raw data is output).

Also, there were problems importing report options from the
request parameters due to some changes to how params are handled
in Rails 5.  I was able to work around that by pulling the
reports into a custom controller and inserting the correct data
via a public method. This may not prove to be the best solution
in the long run.

In summary, it's easy to create and format reports, but it gets
more difficult if it cannot be done in one SQL query.

Steps to Add a Report.

1. Add a new report in `app/reports`.  Enter the sql query in the
   `sql()` method, and the report name in `formatted_title()`.
   Other format method can be seen in the comitted examples.
2. Add a entry in the REPORTS hash in `ReportsController` with
   the name and class initialization in the Proc.
3. If no special formatting is needed for the report, that's it.
   If not, then create a view for it under
   `app/views/dossier/reports` with the correct name and make
   the required edits.

Todo/Next Steps:

* Needs to be able to export in fixed width format for the CNPS
  report.  This is already a card in Trello.
* Determine, by building other reports, if this gem is sufficient
  for our needs.