[Loans/Payments] Handle in Payslips
Also gave payments an explicit date attribute.

Any payments during the period will be added to the payslip
as deductions.  Additionally, any loan balance will be
captured in the payslip object as a record for what was owed
that month.  It will not change as the loan balances change
over time.