Roles and premissions
Users now have roles that can be set by editing a user, this is
also viewable on the users index page. The roles are:
  * Admin
  * Supervisor
  * User

Admins:
  can do all (create, read, update, destroy)
Supervisors:
  cannot create new employees
  can see their own direct reports.
  can edit their direct reports (employee information)
  cannot delete employees
Users:
  can only see their own information and can do nothing else.

Gem for `access-granted` has been added and a policy file
created to manage user roles and the various permissions on the
site.

A preliminary set of permission checks have been added in
`EmployeeController` and unit tests have been created to verify
behavior and ensure that future changes keep access controls
correct. Other controllers still need tests, policies, and
`authorize!` checks.  This is TBD.

NB: The permissions for changing user roles has not been locked
down (yet!).  Anyone can make themselves an administrator.  Once
this is in place, we will need a way to seed proper accounts into
the system on db load (or similar) so that we can still login and
administer the system.

Note: also tidied up the i18n for the newly added admin pages.