Users: Authorization and Permission
Note: One other functional change in this commit, when
a user is submitted, previously the redirect was to the
index page of users.  I didn't give everyone access to
this page, so I just have it now reload the edit page
with a notice that changes were saved.

If a referring url is provided, it will still follow
to that URL after the save regardless (with the notice)

I also made the notice site-wide (in the application
layout).

IMPORTANT: I added an admin user to db/seeds.rb
because now if one is not added, you can't use the
site.  The password should be change after the db
is loaded and seeded.  If you already have an
admin user and are just migrating your database,
this isn't a big deal.