Compute Taxes and BonusBase
This commit adds logic to the Employee, WorkHour,
and Payslip models so that an employee's taxes
and bonusbase can be calculated.  After this commit
caissebase and cnpswage will be added.

Items needed to compute this information:

An employee's `dailyrate` and `hourlyrate` as well
as the number of `hours_per_month` and
`workdays_per_month`

The BonusBase what is used to calculate the Prime
Ancienneté and is one of the values used in
calculating the CNPSWage for taxes.

A new Tax model has been added including its lookup
table imported from the old system. (see REQUIRED
ACTIONS)

Also note, this needs to integrate better with the
earnings model which is already attached to the
payslip. That work will be done in a future commit.

ALSO ADDED:

Employees' `wage_scale` have been modified, which is now
a - e, instead of 1,2,3.  These correspond to different
wages in the wages table (basewage -> basewagee). Wage
lookups in `Wage` and `Employee` have been modified
accordingly.

Also added a boolean that should be set if the employees
spouse is employed.  This is used for some of the tax
calculations to not take taxes from both spouses.

REQUIRED ACTIONS:

You must run `rails db:seed` to get the new
Taxes table.  Likely tests would faill without this.