Bonuses: Now stored as decimals
Bonus quantities for percentage bonuses are still entered
and displayed as numbers between 0 and 100, but are stored
in the database, and therefore used as such in calculations
regarding the bonuses. (see Payslip::process_bonuses)

This is done by using a separate getter/setter that post
processes the value after getting or before setting the
actual `quantity` value in the Bonus model.