Now on Payslip: Wage, Echelon, WageScale, etc.
Added to each payslip record the employee's wage,
echelon, and wagescale information at the time
the payslip was processed, as well as the employee's
monthly totals for:
  * days (days worked this month, if not full month)
  * hours (hours worked this month, if not full month)
  * overtime hours (0-8 hours)
  * overtime2 hours (8-16 hours)
  * overtime3 hours (16+ hours)
  * overtime rate (hourly rate * 1.2)
  * overtime2 rate (hourly rate * 1.3)
  * overtime3 rate (hourly rate * 1.4)

NOTE:  The days or hours columns will be filled in
if the employee is paid monthly or hourly.  If both
are not filled in, then this employee worked a full
month and was not paid for a partial month. Full
wage was applied.