Additional Payslip Fields and Values
This commit adds more fields to the Payslip object, or
fills in existing fields with new values.  These are:

* Gross Pay: Right now this is the same as `taxable`
* Net Pay: Gross Pay less all taxes and deductions.
  This should be what the employee actually gets
  paid
* Total tax: Sum of all the taxes computed by the
  Tax model
* Hourly rate: Stores the computed hourly rate from
  Employee
* Daily rate: Stores the computed daily rate from
  Employee

Other Changes:

Employee: Union dues are now calculated on the `base
wage` not the `wage`.  This is correct per the old
system.

Tax: I refactored how the rounded pay was calculated
a bit and also passed in the `cnpswage` since that is
needed (not the full taxable wage) for computing the
CNPS tax.

TestHelper: Made the employee returned by
`return_valid_employee()` slightly more "complete"
and requiring less configuration in the test.