Percentage Bonuses can have a maximum.
If a percentage bonus has a maximum, the effective bonus
can never be calculated to be above the maximum set on
the model. Thus for the Prime Exceptionnelle (30%) it is
possible for some employees to hit the 55000 CFA limit
with 30% of their CaisseBase, these employees will see
the maximum limit as the bonus, instead of the full 30%.

Fixed amount bonuses cannot have a maximum and have
validation to prevent this.