Refactor and Standardize on Earnings.
Some fields related to earnings were in an Earning
model, and some were on the payslip directly.  This
commit refactors the Payslip processing methods to
use Earning models exclusively for earnings.

This commit also puts into practice using the
`effective_bonus` method introduced into Bonus in
a previous commit to compute the bonus amount (taking
into account any maximum, if applicable).

TODO:

Some values have been left and are still stored on the
payslip for now.  These are related to overtime earnings.
These will be removed in a future commit and will only
appear as associated Earning models on the Payslip