DIPE Text Report
New DIPE report that uses fixed-width output to be
sent and uploaded to the CNPS website.

This report uses a new gem 'fixy' to help with the
structuring of the fixed-width file.

NOTE:

This text file follows documentation from the CNPS.
This was originally found at:

http://www.cnps.cm/language/docs/dipe.pdf

This matches the format of the current dipes text
report produced from the Access EPS.  A copy of the
documentation is now housed at:

\\john\csdept$\eps\CNPS Documentation\dipe.pdf

In case the CNPS document becomes unavailable in the
future

ALSO NOTE:

This commit requires a new table `dipe_codes`.
Therefore you have to run `db:seed` before everything
will work properly to load data into the new table.