Rewrite Reports to Use a Single Query
Previously, I used some of the formatting methods in
the reporting gem to lookup some of the values in
the report output.  This was easier because it made
the queries simpler.  However, when the CSV and XLS
formats are requested, the raw data is returned
(without formatting), and thus data would be odd at
best, or at worst, incorrect.

This is still the case for enum fields.  For example,
the data shown for the field `marital_status` is an
'M' or 'S' on the HTML report (in english), but an
integer value from the enum on the CSV or XLS output.
It's not clear to a user what number means what value.

Additionally, I'm not sure what a good solution is for
that other than perhaps providing a lookup table to
the integer values that are produced next to the CSV
and XLS buttons, or creating enum lookup tables in the
DB.  Both option seem less than ideal.

Moreover, the mis-match between the category and
echelon values (what's in the `wages` table and the
enums in the `employees` table) is proving to be a
challenge.  My potential solution for that is to
create a lookup table to translate enums to the values
in the `wages` table.

That will be addressed in a future commit.