Department Charges Report & Work Loan Percentages
Add the report model and other things required to display
this new report. The Department Charges Report adds a few
new fields to the payslip model and adds another child
model to a payslip.

__New Model: Work Loan Percentages__

This model belongs to a payslip and captures a department and
a percentage.  Each payslip can have many work loan
percentages. The percentage is the number of hours loaned that
month for that department divided by the total number of hours
worked that month by the employee.  Any "leftover" time after
the loans are calculated are "loaned" to the home department.
(Example: 8 hours of time is loaned to Admin, the rest of the
month is a loan to the "home" department because it was not
loaned).

If there are no work loans, a 100% loan is added against the
employee's "home" department.

Having this table makes querying much simpler in the report's
single SQL statement.

_Also note that the model exists exclusively to support this
report (as well as the WorkLoan model itself, which isn't
used anywhere else in the application at this point)_

__New Fields Added to Payslip__

* Department CNPS: computed from `cnpswage`
* Department Crédit Foncier: computed from `taxable`
* Employee Fund: If the `taxable` wage is above a certain
value, a fixed amount is given
* Employee Contribution: This is 0.  The old report would
always compute this value as 0, so I just added a static
0 on the report

__TODO__

Work Loans need to use an actual department selected from
a list of known values.  The fuzzy lookup that exists right
now to try to find a department from the textbox value will
not be sufficient for real users.

The difficulty is how to loan someone to a person and not a
department. The workaround given will be that this person's
account # needs to be added as a department so that user's
can be found in the department list.

This will be done in a following commit.