(Slightly) Improved handling of payslips with negative pay
The strategy is to throw an error so these payslips can be
flagged and looked at. Added validations to produce this error.

Also linkified the payslips output for each payslip and each
employee after processing all payslips for a period.  This
should make it easier to act on various errors.

Additionally, linkified the logo if for extra things to click
on to get sent back to the home page.