Salary Changes Report
View any category and/or echelon change that happend
this month (compared to the values in last month's
payslip).  Present the old and new values as well as
the old and new salaries.

ALSO NOTE

Now you can have a short (or not so short) description
appear at the top of reports by overriding the
`report_description()` method in your report models.