Remove 'term' attribute from Loans
Each loan is given a term by finance when it is created,
however, there are many times when special dispensations are
given during the repayment period that it doesn't really
make sense to track this information. This attribute was
purely information, so not much implementation has changed.

Since finance won't need to use this field, remove it
instead just making it optional.