[Vacations] Handle redirection using `referred_by`.
Since there a multiple ways to enter vacation for an employee
(either being on the employee's page, or on a generic page where
the user selects the employee), remember which page the user
is on -- and send them back there after the operation is complete.

This will help some user confusion by returning them to the same
page they started from.