[Payslip] Adjust pay computation for partial months.
Previously, the monthly pay for a partial month was
computed by taking the total number of days worked
and multiplying that by the daily rate.

However, because of the way the daily rate is
computed for each employee:

  ( Wage / Hours Per Month ) * Hours Per Day

Where `Hours Per Month` is:

  ( Hours Per Day * Days Per Week ) * 52 / 12

And `Hours per Month` is the same each month. This
algorithm comes up with a daily rate that is
"pretty close", but not exact for each month --
but ultimately the same despite the differing
number of working days in each month. Therefore,
the "additive" method used before this commit
wasn't exactly the same as the new "subtractive"
method.

To match exactly, the computation must figure
the number of days *not worked* in the month,
multiply by the daily rate, and reduce the total
pay by that amount.

The computation in `Payslip` and the tests have been
updated (along with the descriptive text in the
payslip).