Allow viewing of paid loan payments
Previously, the interface his the payments from paid loans.
This worked, however, if you mis-entered any information for
a payment that finihsed paying off a loan, it was impossible
to fix via the interface.

This adds a collapsable div which will show you the paid loan
payments if you want to see them, and hide them normally so
they don't clutter up the screen.