Various payroll fixes
There are 2 significant changes in this commit.

1. Round the overtime rates to the nearest CFA. This was
previously using fractional CFA which didn't make sense
and was producing the wrong amounts.

2. Use `bonusbase` when computing percentage-based
bonuses instead of `caissebase` which, when you think
about it, makes a lot more sense.

Also added the employee ID to the debugging output to
make things easier to understand, and added what number
was being used to compute the seniority bonus.

And adjusted tests.