Some bonuses use CaisseBase others use BonusBase
The Prime de Caisse bonuses specifically. Allow for these
to act differently than the other bonuses.

There is now an option to select the base each bonus will
use (the default is bonusbase).  The different prime de
caisse options will need to be selected to use caissebase
after this change is put on production.

Adjusted views, controllers, and tests.