Build interface to edit `Spouse Employed` field
If the spouse is employed or the employee is female,
do not take out communal tax.  There was already a
test for this and functionality in the Tax model,
however there wasn't a way to edit this field for
an employee. This commit adds the edit screens and
the changes in #show.