Partial Months are now days_worked * daily_rate
This effectively reverts commit aacc8c. The algorithm in
aacc8c is only used for missed days transactions and isn't
as common as it appeared with the limited sample set.

Some payslips are still using the subtractive model, mostly
if the employee is just missing a few days, since a 'MD' or
missed days transaction is used in those cases.

I would like all partial months to use the same algorithm
and will discuss this with finance.

This commit also lets payslips be negative up to -4 (within
a 5 CFA rounding error).