Payslip PDF View.
Create a PDF view of the payslip using Prawn and Prawn::Table.
Prawn was already being used for some of the reports that needed
a PDF output.

Most of the changes in this commit (besides the PDF output
itself) are related to the ability to tab and grab certain pieces
of payslip data so it can be put in the correct place. The HTML
view that was created initially showed certain pieces in a
slightly different order and thus additional logic was needed.

This also introduces a new gem to translate integer to roman
numerals for the display of the category and echelon on the
payslip.