Update computation for daily and houly rate
Previously this used a computation based on the average
number of working hours in a month, times the number of
hours an employee worked per day (usually 8).

Now the system will use this computation:

  ((monthly wage * 12) / 15) / days_week

(where days_week is usually 5) to give a new daily
rate.  This is a slightly better approximation than the
previous one.

This commit also standardizes on computing partial
working months by subtracting the days not worked,
rather than adding days worked. There is also only a
single method, previously both adding and subtracting
were used depending on the transaction code (Days
Worked vs. Missed Days).

Hopefully this produces more consistent results and
simplifies the data entry and the calculations. In my
initial testing the differeces are very small (10 to
2000 CFA) from the previous method.