Vacation Pay Revisions
This moves vacation pay out of the Payslip model to be handled by
the Vacation model.

Overall, there are now regular payslips for the working periods,
less any vacations. For the vacation periods, a vacation
"voucher" replaces the old Vacation Payslips. It's a bit like a
payslip, but it is not a payslip.

For instance, take these scenarios:

1) If the employee works the whole month of January except for a
2 day vacation on the 16th and 17th, this user would receive a
normal January payslip less two days. This employee would also
receive a vacation "voucher" for her two days of vacation. This
voucher would compute the CNPS wage from the current period,
store this in the vacation record and also mark it as paid. Then
using the vacation calculation, figure how much vacation should
be paid for those two days (less any taxes) and produce a
voucher.

2) The employee works part of January, and then takes vacation
until March 3rd. In this case, the employee would receive a
normal payslip for January. This employee would not receive a
February payslip (or a payslip that has zero pay), and receieves
a partial payslip for March.  The vacation voucher would count
all of the days taken between the start of the vacation in
January and the end of the vacation in March and work out the
vacation pay needed (less taxes).

The vacation "Voucher" can be printed like a normal payslip and
be presented to finance for payment.  In this regard, it is like
a normal vacation payslip from the old system.  But it is not a
payslip.

All necessary data is now stored alongside a vacation record.
This commit adds columns for:

  * how much vacation pay the employee should receive for this
    vacation
  * whether a voucher has already been printed (which prevents a
    vacation from being edited or deleted, since, well, you can't
    get paid for something and then undo it)

Vacation now can compute all the necessary taxes (really this
just uses the Tax model as-is with the `vacation_pay` stored in
the vacation record).  Tax data is not persisted but it can be
re-computed at runtime.

TODO/Next Steps:

[ ] Verify that vacation balances are used up correctly.
[ ] Supplementary days. How work?
[ ] Revise voucher display to present a PDF instead (or in
    addition to) the HTML view.
[ ] What happens during months when no work is done (because
    on vacation the entire time? Is this no payslip, or a
    zero dollar payslip?