Standardize on subtracative for partial month pay.
After talking to Hilda, she thinks this is the best method, which
the code was already using.  However, I have changed the daily rate
computation back to the original used in the old system. This is
going to be evaluated by Hilda once she can generate payslips on
the production system.