Fix Vacations Not Clobbering Work Hours
If a vacation is added, any work hours over that same
period should be removed. This wasn't working properly
because the controller was calling `save!` and `save`
was overridden in the method.  Since there are multiple
ways to add an associated object and `save!` and `save`
are both used, it only seems appropriate to add the
required behavior to both.