Implement New Report using ThinReports.
[ThinReports](https://github.com/thinreports) is a reporting
gem that required an external editor to edit a report file.
The report file contains text fields that can be substituted
at runtime.

The external editors is a GUI app that produces JSON output
describing the report's layout.

The advantage over the other gems used in reporting (which
there are many, unfortunately) like Dossier or pure Prawn is
that there is less effort in forcing summary rows into the
query results (in the case of dossier) or mixing in all the
formatting and layout items into the report itself (in the
case of pure Prawn).

Since the layout of the report can be defined separately it
does allow for better separation of concerns, and ideally a
cleaner reporting file. Although that is not the case.

In this case, the DIPES internal report was changed from a
pure Dossier report to a ThinReports report. Some changes
were required in the Controller to allow for the ThinReports
gem to handle processing. Most notably a key for `format`
was added to the hash of available reports indicating that
this report prefers to be displayed in PDF.

The other PDF reports will be migrated in future commits and
hopefully the fixed width text-only report can follow this
pattern.

One of the downsides is that the generation is a bit slow,
but this could be how I'm implementing things right now and
could be improved in the future.