[Vacation Overhaul]
Based on other revisions to the reports and getting the data
to line up correctly with previous information required a
revision of how vacations workekd.

NOTE: An associated file which is not in version control
needs to be applied when the new fields are added to give
all employees their previous month's vacation balances and
for the counting to begin correctly. This is called
`vac_updates_dec.txt` and needs to be loaded manually into
each employee's December payslip.

Basic overview of the changes:

1) Accumulated Days

Vacation could not simply be computed based on the current
pay, if the employee earned less over previous months, the
accrued vacation pay would be lower those months, and similarly
if a bonus is received. This change uses the accumulated
vacation days and pay to compute a daily rate (based on the
accumulated days) and this is awarded over the span of the
vacation.

2) Supplemental days.

Supplemental days are largely unchanged, except that when
they are awarded (on the same month as your contract start
date, anually) the accumulation is cleared and reset.

3) Reports Updates

For the `Department Charge Report`: update report based on
the new fields, additionally all of the vacation numbers
need to handle the yet-to-be-given supplemental days in the
calculations (they do this by computing the accumulated rate
and applying it by the yet-to-be-given days count). This makes
the query a bit more complex.