Revisions to Work Loan Percentages calculations.
Computer loans based on a "standard month", not on the actual
number of working hours per month. This means January 2018,
which has 23 working days and thus 184 working hours in a 40
hour week, should use 173.33 hours.  This is the standard
number of hours per month, or more accurately, the average
number of hours per month. This is used for the percentage
calculation.

The only exception is if the month worked is not a complete
month, then the number of days * the hours per day is used
for the calculation. If I am off 4 days in January 2018, then
I will use `loaned hours / ((23 - 4) * 8)` as the calculation.

Updated tests as appropriate.