Stop Double Counting Supplemental Days
The previous commit changed application of supplemental days,
so this prevents them from being added by vacation and by
payslip.

However, this does leave a case where supplemental days are
not applied at all if they don't need to be.

It needs to figure out what algorithm is going to be used
and what the requirements are: i.e. A way to request transfering
days?

Also: fixed a bug with holidays and working days, which is
included in this commit.