Supplementary Days Updates
After these changes, supplemental days rely a bit less on the
database to keep running totals, and can now keep track somewhat
automatically as long as the DB knows when the last transfer happened.

Add a Supplementary Transfer model to handle the history of when
transfers happened for each employee.

Supplementary Transfers are counted from the most recent transfer
that isn't happening in the current period.

A question about supplementary transfer is if they are just leftovers
from the "old way" of doing things, but the end is in sight.