Add Holiday/Vacation Hours Worked
Theses are hours an employee worked during a vacation that
the employee should get paid for. This is on top of the
vacation pay that should be received. These are paid at the
regular hourly rate.

Validations have been updated. Vacation hours cannot be
taken on days where the employee is not on vacation, and
regular hours cannot be taken during a vacation.

One question that still needs to be answered is if vacation
hours worked on a Sunday should be paid at the holiday overtime
rate? Or is it always paid at the normal rate? Things get
complicated.

This has been updated in the payslip processing, this is a
special kind of overtime. Also updated the tests.