[Payslips] handle inactive gaps
Some employees spend a month or more inactive, to
suddenly return. This modifies the existing behavior
to better handle when there are gaps in the payslip.

This improves the logic of finding the previous
payslip. Previously, the code assumed this would
always be in the previous period. Now it attempts to
find the last one by query.

Also, now all employees are processed during payslip
processing (i.e. post period, or process all payslips
for a period). If there is no need to create a payslip
it simply moves on. If a payslip was previously
created and the employee is now inactive for this
period, the payslip will be _deleted_.

There still is no way to track periods where an
employee is inactive, so this still must be done by
marking the user inactive and active at the correct
points manually.