Category and Echelon Fixed with Payslip
Previously, a payslip would always show the current
category and echelon set for an employee. Thus if the
employee got a raise, this would change the display
on the older payslips before the raise occured.  Now
the change on older payslips shows what category and
echelon was set for the employee at the time of the
payslip.

This simplifies the display of the category and echelon
across the application and moves the display function
of each into the `application_helper`. The application
helper has a function for display of the category as a
number and as a roman numeral.

Other associated changed as needed.