Fix Child Tests That Weren't Locked in Time
These broke in 2019 because the children kept getting
older and eventually they were older than 19, thus
breaking the test. This change fixes the test in time,
preventing the children from getting older and making
sure the test always passes. This moves the `now()`
date from the database, to something that comes from
Rails, which can be overridden by a method stub.

This also fixes a minor typo in a comment.