Add Minimum to Bonus
Add a minimum amount value for bonuses that are percentage-based.
In the case of a bonus with a minimum, if the effective bonus
computed for an employee is lower than the minimum, then the
employee will receive the minimum amount and no less. This is
similar to the bonus maximum which already existed, where if the
effective bonus is greater than the maximum, the employee will
receive the maximum amount and no more.

Note that it is possible to have both a minimum and a maximum in
a bonus.

This is for the new logement bonus which is percentage based,
but everyone will receive at least 7 500 CFA (or something like
that).