Handle category/echelon in voucher if no payslip.
Since there's nothing to look up, this would throw
an error. Fall back to the employee's value, although
the category/echelon may have changed since this vacation
period (although if it's in the past, there should be a
payslip found by the query) -- this should be only for
period where the vacation voucher is viewed before there
is a payslip yet.