Fix double counting of union dues
After previous commit about union dues, this caused a double
counting because it was added to total tax, but it still existed
as a deduction. Thus, it was counted against the gross pay
twice. This makes union_dues an actual field on the payslip
and thus can be counted only once (part of tax, no longer a
deduction).