Add "Salaire Net" to the Dept Charge Report as Net Pay
This value should be the net pay (after tax) but before
deduction. This is labeled on the payslip as "Salaire Net
(arrondi)". This value was calculated via a method.
However, to add it to the report, it needs to be a
database field that can be queried -- thus it get a
migration, etc., and all associated changes.