Vacation Pay and Taxes and Reports
Persist the vacation pay tax information so that it can
be queried in the reports.

Also modify the report queries to include any vacation
information (for the DIPES Internal, DIPES Goverment, and
DIPES Text reports) in the pay and tax lines for each
employee. These are combined lines -- not separate lines
for each payslip as in the old system reports. This is
okay per Hilda.

This commit also eliminates "magic numbers" in the reports
and replaces them with system variabled, either using an
existing variable, or creating a new variable were one
did not exist before.