Misc Payments require choice of Before/After Tax
Before after tax was assumed. If you did nothing, the payment
would be after tax. No, a choice must be made for either Before
or After. If nothing is done, an error is thrown.

Note: for some reason I couldn't get the translations for Yes or
No working, so I just put Yes and No in the template. I have
no idea why.