Minor Updates to the Department Charges Report.
(Large update to the Charge and Deduction models)

Salaire Net needs to be Net Wage less Advances and Taxes.

Thus:
  * Create new type in Charge (and thus Deduction) for
    various charge types. This is now easily queryable
    despite what the note says.

These need to updated appropriately on live after the
migration runs. The plan is to set all charges and
deductions to ID 2, then decide one by one for the
others.