Updates to the DIPES reports
Mostly: make the text report look like the example.

Note: These changes mostly anticipate a coming change in which
an employee may not receive a payslip in a given month (if, for
example, they were on vacation the whole month). This report
combines possibly non-existent payslips with possibly non-exitent
vacation pay in one line.

In other reports, the user should not appear at all if they do not
receive a playslip.