Not working during the month makes you get 0
This is mostly likely encountered if you are on
vacation for an entire month (not that unusual).
In this case, you would have received a vacation
voucher for the month to receive your vacation pay
and thus would not receive a normal regular payslip.
You receive a payslip, but it just contains zero pay.

Also note that vacation does not accrue when you are
on vacation the whole month.

This change makes all the reports correct if you are
vacation the whole month as well (mostly the department
charge report).

This change also breaks away from the Vacation.days_used
calculation which wasn't based on the historical payslip
records. This change uses the last payslip to figure out
it's balances regarding supplemental days.

Further refactoring is needed to complete what was
started.