340 Journal Report Page
Make an "easy-to-follow" (maybe) 340 journal page that simplifies
entry into the monthly payroll journal. Eventually this could be
used to generate an XLS of the journal itself.

JOURNAL OVERVIEW

Basically this determines all the money paid to employees (DEBITS)
through Department Budgets and displays the accounts the are given
that money (CREDITS) based on where the money goes (i.e. Cash paid
from Finance, taxes, etc.)

NOTES

By default the report shows the Last Posted Period, but periods
are selectable so a user can see a previous period as needed.

FUTURE WORK NEEDED

* Departments currently have one single account number associated
  with them. However, they need to be able to have multiple
  cost centers (Payroll Charges, Transportation, Taxi, etc.) this
  is ignored at this point.
* Perhaps a better implementation is to determine a list of all
  items for the report, and then pop them from the list as they
  are displayed, leaving anything unaccounted for.
* Better internal lookup of the various account numbers and names
  instead of just being littered throughout the report.