Transfer This Period's Supplemental Days As Well
During transfer, the current period's supplemental days
was not included in the amount transfered.

For example, if the April supplemental balance was 4, and
the employee earned 0.33 supplemental days per period, when
the transfer happened in May, only 4 days would be
transferred. This fixes this amount to 4.33 to include the
current period's amount.

This amount is not included if they are on vacation the
entire period (since you don't earn if that happens).

Tests have been updated accordingly.